pyproject.toml
[tool.poetry]
name = "event_scheduler"
version = "1.0.0"
description = "REST API for a tool which helps to organize meetings."
authors = ["Aidar Garikhanov <a1d4r@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = {extras = ["standart"], version = "^0.13.4"}
intervaltree = "^3.1.0"
tortoise-orm = "^0.17.2"
aerich = "^0.5.3"
asyncpg = "^0.22.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
autoflake = "^1.4"
isort = "^5.7.0"
black = "^20.8b1"
unify = "^0.5"
flake8 = "^3.8.4"
pylint = "^2.7.1"
mypy = "^0.812"
pytest-deadfixtures = "^2.2.1"
asynctest = "^0.13.0"
pytest-asyncio = "^0.15.1"
httpx = "^0.17.1"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
